+++ src/core/operators/switch.ml
@@ -157,8 +157,8 @@ class switch ~all_predicates ~override_meta ~transition_length ~replay_meta
                        * A quicker hack might have been doable if there wasn't a
                        * transition in between. *)
                         match c.source#last_metadata with
-                        | Some m when replay_meta ->
-                            new Insert_metadata.replay m c.source
+                        | Some (_, m) when replay_meta ->
+                            new Replay_metadata.replay m c.source
                         | _ -> c.source
                     in
                     Typing.(new_source#frame_type <: self#frame_type);
@@ -188,8 +188,8 @@ class switch ~all_predicates ~override_meta ~transition_length ~replay_meta
                        * A quicker hack might have been doable if there wasn't a
                        * transition in between. *)
                         match c.source#last_metadata with
-                        | Some m when replay_meta ->
-                            new Insert_metadata.replay m c.source
+                        | Some (_, m) when replay_meta ->
+                            new Replay_metadata.replay m c.source
                         | _ -> c.source
                     in
                     Typing.(old_source#frame_type <: self#frame_type);