+++ src/core/outputs/icecast2.ml
@@ -533,7 +533,7 @@ class output p =
         | Cry.Connected _, Some enc -> enc.Encoder.encode frame
         | _ -> Strings.empty
 
-    method insert_metadata m =
+    method encode_metadata m =
       (* Update metadata using ICY if told to.. *)
       if send_icy_metadata then (
         let f = Charset.convert ~target:out_enc in
@@ -564,7 +564,7 @@ class output p =
       else (
         (* Encoder is not always present.. *)
           match encoder with
-          | Some encoder -> encoder.Encoder.insert_metadata m
+          | Some encoder -> encoder.Encoder.encode_metadata m
           | None -> ())
 
     method icecast_send b =
@@ -643,9 +643,9 @@ class output p =
         self#log#important "Connection setup was successful.";
 
         (match source#last_metadata with
-          | Some m when send_last_metadata_on_connect -> (
+          | Some (_, m) when send_last_metadata_on_connect -> (
               try
-                self#insert_metadata
+                self#encode_metadata
                   (Frame.Metadata.Export.from_metadata ~cover:false m)
               with _ -> ())
           | _ -> ());