+++ src/lang/doc.ml
@@ -191,7 +191,13 @@ module Value = struct
   }
 
   let db = ref Map.empty
-  let add (name : string) (doc : t Lazy.t) = db := Map.add name doc !db
+
+  let add (name : string) (doc : t Lazy.t) =
+    let name =
+      Re.replace ~all:true ~f:(fun _ -> "null") (Re.Pcre.regexp "^_null") name
+    in
+    db := Map.add name doc !db
+
   let get name = Lazy.force (Map.find name !db)
   let count () = Map.cardinal !db