+++ tests/core/output_encoded_test.ml
@@ -1,4 +1,4 @@
-let insert_metadata_called = ref false
+let encode_metadata_called = ref false
 let encode_called = ref false
 let send_called = ref false
 
@@ -17,9 +17,9 @@ class encoded_test =
 
     method self_sync = (`Static, None)
 
-    method insert_metadata _ =
+    method encode_metadata _ =
       assert !send_called;
-      insert_metadata_called := true
+      encode_metadata_called := true
 
     method encode _ =
       encode_called := true;
@@ -44,4 +44,4 @@ let () =
   let m = Frame.Metadata.from_list [("foo", "bla")] in
   let frame = Frame.add_metadata frame 0 m in
   encoded_test#test_send_frame frame;
-  assert !insert_metadata_called
+  assert !encode_metadata_called